+++ packages/server/database/migrations/20201224143827_add-roadmap-id-to-posts-table.js
@@ -1,9 +1,9 @@
 // utils
 const logger = require("../../utils/logger");
 
-exports.up = (knex) => {
-  return knex.schema
-    .table("posts", (table) => {
+exports.up = async (knex) => {
+  if (!(await knex.schema.hasColumn('posts', 'roadmap_id'))) { // Check if column exists
+    return knex.schema.table("posts", (table) => {
       table
         .uuid("roadmap_id")
         .references("id")
@@ -13,29 +13,37 @@ exports.up = (knex) => {
     .then(() => {
       logger.info({
         code: "DATABASE_MIGRATIONS",
-        message: "Add column: roadmap_id in posts",
+        message: "Adding column: roadmap_id in posts",
       });
     })
     .catch((err) => {
       logger.error({
-        message: err,
+        code: "DATABASE_MIGRATIONS",
+        message: err.message, // Log the actual error message for better debugging
       });
     });
+  }
 };
 
 exports.down = (knex) => {
   return knex.schema
-    .table("posts", (table) => {
-      table.dropColumn("roadmap_id");
+    .hasColumn("posts", "roadmap_id")
+    .then((exists) => {
+      if (exists) {
+        return knex.schema.table("posts", (table) => {
+          table.dropColumn("roadmap_id");
+        });
+      }
     })
     .then(() => {
       logger.info({
-        message: "Drop column: roadmap_id in posts",
+        message: "Dropping column: roadmap_id in posts",
       });
     })
     .catch((err) => {
       logger.error({
-        message: err,
+        code: "DATABASE_MIGRATIONS",
+        message: err.message, // Log the actual error message for better debugging
       });
     });
 };
