+++ packages/server/database/migrations/20210626162044_insert-labs-post_activity.js
@@ -1,37 +1,50 @@
 // utils
 const logger = require("../../utils/logger");
 
-exports.up = (knex) => {
-  return knex.schema
-    .table("settings", (table) => {
-      table.text("labs").defaultTo("{}");
-    })
-    .then(() => {
-      logger.info({
-        code: "DATABASE_MIGRATIONS",
-        message: "Insert column 'labs' into settings",
+exports.up = async (knex) => {
+  if (!(await knex.schema.hasColumn('settings', 'labs'))) { // Check if column exists
+    return knex.schema
+      .table("settings", (table) => {
+        table.text("labs").defaultTo("{}");
+      })
+      .then(() => {
+        logger.info({
+          code: "DATABASE_MIGRATIONS",
+          message: "Adding column 'labs' to settings",
+        });
+        return knex("settings").update({
+          labs: knex.raw("labs::jsonb || '{\"comments\": false}'"),
+        });
+      })
+      .then(() => {
+        logger.info({
+          code: "DATABASE_MIGRATIONS",
+          message: "Inserting data into 'labs' column",
+        });
+      })
+      .catch((err) => {
+        logger.error({
+          code: "DATABASE_MIGRATIONS",
+          message: err.message,
+        });
       });
-    })
-    .then(() => {
-      return knex("settings").update({
-        labs: knex.raw("labs::jsonb || '{\"comments\": false}'"),
-      });
-    })
-    .then(() => {
-      logger.info({
-        code: "DATABASE_MIGRATIONS",
-        message: "Insert 'comments' into labs column",
-      });
-    })
-    .catch((err) => {
-      logger.error(err);
+  } else {
+    logger.warn({
+      code: "DATABASE_MIGRATIONS",
+      message: "'labs' column already exists in settings table. Skipping migration.",
     });
+  }
 };
 
 exports.down = (knex) => {
   return knex.schema
-    .table("settings", (table) => {
-      table.dropColumn("labs");
+    .hasColumn("settings", "labs")
+    .then((exists) => {
+      if (exists) {
+        return knex.schema.table("settings", (table) => {
+          table.dropColumn("labs");
+        });
+      }
     })
     .then(() => {
       logger.info({
@@ -39,6 +52,10 @@ exports.down = (knex) => {
       });
     })
     .catch((err) => {
-      logger.error(err);
+      logger.error({
+        code: "DATABASE_MIGRATIONS",
+        message: err.message,
+      });
     });
 };
+
